+++ pylamarzocco/util/_authentication.py
@@ -1,4 +1,5 @@
-"""Helpers for the pylamarzocco package."""
+"""Authentication related utilities."""
+
 
 import base64
 import hashlib
@@ -6,7 +7,6 @@ import time
 import uuid
 from dataclasses import dataclass, field
 
-from aiohttp import ClientResponse
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric.ec import (
     ECDSA,
@@ -17,71 +17,10 @@ from cryptography.hazmat.primitives.asymmetric.ec import (
 from mashumaro import field_options
 from mashumaro.mixins.json import DataClassJSONMixin
 
-from .const import StompMessageType
-
-
-def is_success(response: ClientResponse) -> bool:
-    """Check if response is successful."""
-    return 200 <= response.status < 300
-
-
-def encode_stomp_ws_message(
-    msg_type: StompMessageType, headers: dict[str, str], body: str | None = None
-) -> str:
-    """Encode STOMP WebSocket message."""
-    fragments: list[str] = []
-    fragments.append(str(msg_type))
-    for key, value in headers.items():
-        fragments.append(f"{key}:{value}")
-    msg = "\n".join(fragments)
-    msg += "\n\n"
-    if body:
-        msg += body
-    msg += "\x00"
-    return msg
-
-
-def decode_stomp_ws_message(
-    msg: str,
-) -> tuple[StompMessageType, dict[str, str], str | None]:
-    """Decode STOMP WebSocket message."""
-    header, data = msg.split("\n\n", 1)
-    headers: dict[str, str] = {}
-    metadata = header.split("\n")
-    msg_type = StompMessageType(metadata[0])
-    for header in metadata[1:]:
-        key, value = header.split(":", 1)
-        headers[key] = value
-    if data.endswith("\x00"):
-        data = data[:-1]
-    return msg_type, headers, data
-
 
 def b64(data: bytes) -> str:
     """Base64 encode bytes to ASCII string."""
     return base64.b64encode(data).decode("ascii")
-
-
-def generate_request_proof(base_string: str, secret32: bytes) -> str:
-    """La Marzocco's custom proof generation algorithm (Y5.e equivalent)"""
-    if len(secret32) != 32:
-        raise ValueError("secret must be 32 bytes")
-
-    work = bytearray(secret32)  # Make mutable copy
-
-    for byte_val in base_string.encode("utf-8"):
-        idx = byte_val % 32
-        shift_idx = (idx + 1) % 32
-        shift_amount = work[shift_idx] & 7  # 0-7 bit shift
-
-        # XOR then rotate left
-        xor_result = byte_val ^ work[idx]
-        rotated = (
-            (xor_result << shift_amount) | (xor_result >> (8 - shift_amount))
-        ) & 0xFF
-        work[idx] = rotated
-
-    return b64(hashlib.sha256(work).digest())
 
 
 @dataclass
@@ -125,6 +64,27 @@ class InstallationKey(DataClassJSONMixin):
             hashlib.sha256(pub_bytes).digest()
         )  # Hash raw bytes, not base64 string
         return f"{self.installation_id}.{pub_hash_b64}"
+
+def generate_request_proof(base_string: str, secret32: bytes) -> str:
+    """La Marzocco's custom proof generation algorithm (Y5.e equivalent)"""
+    if len(secret32) != 32:
+        raise ValueError("secret must be 32 bytes")
+
+    work = bytearray(secret32)  # Make mutable copy
+
+    for byte_val in base_string.encode("utf-8"):
+        idx = byte_val % 32
+        shift_idx = (idx + 1) % 32
+        shift_amount = work[shift_idx] & 7  # 0-7 bit shift
+
+        # XOR then rotate left
+        xor_result = byte_val ^ work[idx]
+        rotated = (
+            (xor_result << shift_amount) | (xor_result >> (8 - shift_amount))
+        ) & 0xFF
+        work[idx] = rotated
+
+    return b64(hashlib.sha256(work).digest())
 
 
 def generate_extra_request_headers(installation_key: InstallationKey) -> dict[str, str]:
